Remove `Ord` impl for all decimal types
while using this crate I found it confusing that a decimal type had `Ord` implemented, as it's unclear to me what comparing `NaN` to some other value would result in.

For example both `f64` and `f32` do not implement `Ord` but just `PartialOrd`. I think this is possibly more correct.